requirements.txt
streamlit
pandas
playwright
pytest
pytest-playwright
google-generativeai
rdflib
pyyaml
pyshacl 